HH83: update Header and Navbar
