+++ src/components/ui/index.ts
@@ -1,1 +1,3 @@
 export * from './AuthLogo';
+export * from './Header';
+export * from './Navbar';
